project/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-countries==7.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==1.7.1
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.4.0
